DCS-1269: Process recall - re-activate most recent booking
