+++ src/main/kotlin/uk/gov/justice/digital/hmpps/welcometoprison/model/arrival/ImprisonmentStatus.kt
@@ -1,0 +1,20 @@
+package uk.gov.justice.digital.hmpps.welcometoprison.model.arrival
+
+import com.fasterxml.jackson.annotation.JsonInclude
+import io.swagger.v3.oas.annotations.media.Schema
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@Schema(description = "The imprisonment status")
+data class ImprisonmentStatus(
+  @Schema(description = "The imprisonment status", example = "Determinate sentence")
+  val description: String,
+
+  @Schema(description = "Associated Nomis code", example = "SENT")
+  val imprisonmentStatusCode: String,
+
+  @Schema(description = "Title for Movement reasons page, (if required)", example = "What is the type of determinate sentence?")
+  val secondLevelTitle: String? = null,
+
+  @Schema(description = "Movement reasons data", example = "Intermittent custodial sentence, INTER ")
+  val movementReasons: List<MovementReason>,
+)
