+++ src/main/kotlin/uk/gov/justice/digital/hmpps/welcometoprison/model/arrival/MovementReason.kt
@@ -1,0 +1,14 @@
+package uk.gov.justice.digital.hmpps.welcometoprison.model.arrival
+
+import com.fasterxml.jackson.annotation.JsonInclude
+import io.swagger.v3.oas.annotations.media.Schema
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@Schema(description = "The reason for a movement into prison")
+data class MovementReason(
+  @Schema(description = "Reason for movement, (if required)", example = "Intermittent custodial sentence")
+  val description: String? = null,
+
+  @Schema(description = "Associated Nomis code", example = "INTER")
+  val movementReasonCode: String,
+)
