+++ src/test/kotlin/uk/gov/justice/digital/hmpps/welcometoprison/model/arrival/ArrivalsServiceConfirmArrivalTest.kt
@@ -8,7 +8,7 @@ import org.assertj.core.api.Assertions
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.ValueSource
+import org.junit.jupiter.params.provider.MethodSource
 import uk.gov.justice.digital.hmpps.welcometoprison.model.basm.BasmService
 import uk.gov.justice.digital.hmpps.welcometoprison.model.confirmedarrival.ArrivalType
 import uk.gov.justice.digital.hmpps.welcometoprison.model.confirmedarrival.ConfirmedArrivalService
@@ -22,6 +22,7 @@ import java.time.Instant
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.ZoneId
+import java.util.stream.Stream
 
 class ArrivalsServiceConfirmArrivalTest {
   private val prisonService: PrisonService = mockk(relaxUnitFun = true)
@@ -29,26 +30,26 @@ class ArrivalsServiceConfirmArrivalTest {
   private val prisonerSearchService: PrisonerSearchService = mockk()
   private val confirmedArrivalService: ConfirmedArrivalService = mockk(relaxUnitFun = true)
 
-  private val arrivalsService =
-    ArrivalsService(basmService, prisonService, prisonerSearchService, confirmedArrivalService, FIXED_CLOCK)
-  val result = { prisonNumber: String?, pnc: String? -> MatchPrisonerResponse(prisonNumber, pnc, "ACTIVE IN") }
+  private val arrivalsService = ArrivalsService(
+    basmService, prisonService, prisonerSearchService, confirmedArrivalService, FIXED_CLOCK
+  )
 
   @Test
   fun `Confirm arrival not known to NOMIS`() {
 
-    every { basmService.getArrival(any()) } returns prototypeArrival.copy()
+    every { basmService.getArrival(any()) } returns ARRIVAL_PROTOTYPE.copy()
     every { prisonerSearchService.getCandidateMatches(any()) } returns emptyList()
     every { prisonService.createOffender(any()) } returns OFFENDER_NO
     every { prisonService.admitOffenderOnNewBooking(any(), any()) } returns BOOKING_ID
 
-    val response = arrivalsService.confirmArrival(MOVE_ID, confirmArrivalDetail)
+    val response = arrivalsService.confirmArrival(MOVE_ID, CONFIRMED_ARRIVAL_DETAIL_PROTOTYPE)
 
     assertThat(response.offenderNo).isEqualTo(OFFENDER_NO)
 
     verify { basmService.getArrival(MOVE_ID) }
     verifySequence {
-      prisonService.createOffender(confirmArrivalDetail)
-      prisonService.admitOffenderOnNewBooking(OFFENDER_NO, confirmArrivalDetail)
+      prisonService.createOffender(CONFIRMED_ARRIVAL_DETAIL_PROTOTYPE)
+      prisonService.admitOffenderOnNewBooking(OFFENDER_NO, CONFIRMED_ARRIVAL_DETAIL_PROTOTYPE)
     }
     verify {
       confirmedArrivalService.add(
@@ -63,110 +64,102 @@ class ArrivalsServiceConfirmArrivalTest {
   }
 
   @Test
-  fun `Confirm arrival matched to NOMIS offender who is not in custody`() {
-
-    every { basmService.getArrival(any()) } returns prototypeArrival.copy(
-      prisonNumber = OFFENDER_NO,
-      isCurrentPrisoner = false
+  fun `Confirm arrival matched to NOMIS offender who is not in custody - with booking in time`() {
+    doParameterizedConfirmArrivalTest(
+      "",
+      ArrivalType.NEW_BOOKING_EXISTING_OFFENDER,
+      BOOKING_IN_TIME,
+      prisonService::admitOffenderOnNewBooking
     )
-    every { prisonerSearchService.getCandidateMatches(any()) } returns listOf(
-      MatchPrisonerResponse(
-        prisonerNumber = OFFENDER_NO,
-        pncNumber = null,
-        status = INACTIVE_OUT
-      )
-    )
-
-    every { prisonService.admitOffenderOnNewBooking(any(), any()) } returns BOOKING_ID
-
-    val response = arrivalsService.confirmArrival(MOVE_ID, confirmArrivalDetail)
-
-    assertThat(response.offenderNo).isEqualTo(OFFENDER_NO)
-
-    verify { basmService.getArrival(MOVE_ID) }
-    verify { prisonService.admitOffenderOnNewBooking(OFFENDER_NO, confirmArrivalDetail) }
-    verify {
-      confirmedArrivalService.add(
-        MOVE_ID,
-        OFFENDER_NO,
-        PRISON_ID,
-        BOOKING_ID,
-        BOOKING_IN_TIME.toLocalDate(),
-        ArrivalType.NEW_BOOKING_EXISTING_OFFENDER
-      )
-    }
   }
 
   @Test
-  fun `Confirm arrival matched to NOMIS offender who is in custody`() {
+  fun `Confirm arrival matched to NOMIS offender who is in custody is rejected`() {
 
-    every { basmService.getArrival(any()) } returns prototypeArrival.copy(prisonNumber = OFFENDER_NO)
+    every { basmService.getArrival(any()) } returns ARRIVAL_PROTOTYPE.copy(prisonNumber = OFFENDER_NO)
     every { prisonerSearchService.getCandidateMatches(any()) } returns listOf(
       MatchPrisonerResponse(prisonerNumber = OFFENDER_NO, pncNumber = null, status = "ACTIVE IN")
     )
 
     Assertions.assertThatThrownBy {
-      arrivalsService.confirmArrival(MOVE_ID, confirmArrivalDetail)
+      arrivalsService.confirmArrival(MOVE_ID, CONFIRMED_ARRIVAL_DETAIL_PROTOTYPE)
     }.isInstanceOf(IllegalArgumentException::class.java)
 
     verify { basmService.getArrival(MOVE_ID) }
   }
 
   @Test
-  fun `Confirm arrival matched to NOMIS offender who is in custody - default arrival time`() {
+  fun `Confirm arrival matched to NOMIS offender who is in custody is rejected - default arrival time`() {
 
-    every { basmService.getArrival(any()) } returns prototypeArrival.copy(prisonNumber = OFFENDER_NO)
+    every { basmService.getArrival(any()) } returns ARRIVAL_PROTOTYPE.copy(prisonNumber = OFFENDER_NO)
     every { prisonerSearchService.getCandidateMatches(any()) } returns listOf(
       MatchPrisonerResponse(prisonerNumber = OFFENDER_NO, pncNumber = null, status = "ACTIVE IN")
     )
 
     Assertions.assertThatThrownBy {
-      arrivalsService.confirmArrival(MOVE_ID, confirmArrivalDetail.copy(bookingInTime = LocalDateTime.now(FIXED_CLOCK)))
+      arrivalsService.confirmArrival(
+        MOVE_ID,
+        CONFIRMED_ARRIVAL_DETAIL_PROTOTYPE.copy(bookingInTime = LocalDateTime.now(FIXED_CLOCK))
+      )
     }.isInstanceOf(IllegalArgumentException::class.java)
 
     verify { basmService.getArrival(MOVE_ID) }
   }
 
   @ParameterizedTest
-  @ValueSource(strings = ["ETRB", "Y", "ETRLR", "ETRRIE", "ETB", "B", "H", "24", "L"])
-  fun `Offender should be recalled when specified by movementReasonCode`(movementReasonCode: String) {
-    every { basmService.getArrival(any()) } returns prototypeArrival.copy(
-      prisonNumber = OFFENDER_NO
+  @MethodSource("recallMovementReasonCodes")
+  fun `Person having a Prison Number should be recalled when specified by movementReasonCode`(movementReasonCode: String) {
+    doParameterizedConfirmArrivalTest(
+      movementReasonCode,
+      ArrivalType.RECALL,
+      null,
+      prisonService::recallOffender
     )
-    every { prisonService.recallOffender(any(), any()) } returns BOOKING_ID
-    every { prisonerSearchService.getCandidateMatches(any()) } returns listOf(
-      MatchPrisonerResponse(prisonerNumber = OFFENDER_NO, pncNumber = null, status = "INACTIVE OUT")
-    )
-    arrivalsService.confirmArrival(
-      MOVE_ID,
-      confirmArrivalDetail.copy(
-        bookingInTime = LocalDateTime.now(FIXED_CLOCK),
-        movementReasonCode = movementReasonCode
-      )
-    )
-    verify { prisonService.recallOffender(any(), any()) }
-    verify { confirmedArrivalService.add(any(), any(), any(), any(), any(), ArrivalType.RECALL) }
   }
 
-  @Test
-  fun `Exising offender with no booking should be admitted on a new booking`() {
+  @ParameterizedTest
+  @MethodSource("nonRecallMovementReasonCodes")
+  fun `Person having a Prison Number should be admitted on a new booking when not recalled`(movementReasonCode: String) {
+    doParameterizedConfirmArrivalTest(
+      movementReasonCode,
+      ArrivalType.NEW_BOOKING_EXISTING_OFFENDER,
+      null,
+      prisonService::admitOffenderOnNewBooking
+    )
+  }
 
-    every { basmService.getArrival(any()) } returns prototypeArrival.copy(
-      prisonNumber = OFFENDER_NO
-    )
-    every { prisonService.admitOffenderOnNewBooking(any(), any()) } returns 1
+  private fun doParameterizedConfirmArrivalTest(
+    movementReasonCode: String,
+    expectedArrivalType: ArrivalType,
+    bookingInTime: LocalDateTime?,
+    expectedOperation: (String, ConfirmArrivalDetail) -> Long
+  ) {
+    every { basmService.getArrival(any()) } returns ARRIVAL_PROTOTYPE.copy(prisonNumber = OFFENDER_NO)
     every { prisonerSearchService.getCandidateMatches(any()) } returns listOf(
-      MatchPrisonerResponse(prisonerNumber = OFFENDER_NO, pncNumber = null, status = "INACTIVE OUT")
+      MatchPrisonerResponse(prisonerNumber = OFFENDER_NO, pncNumber = null, status = INACTIVE_OUT)
     )
-    arrivalsService.confirmArrival(
-      MOVE_ID,
-      confirmArrivalDetail.copy(
-        bookingInTime = LocalDateTime.now(FIXED_CLOCK),
-        movementReasonCode = "TXF"
+    every { expectedOperation(any(), any()) } returns BOOKING_ID
+
+    val confirmArrivalDetail = CONFIRMED_ARRIVAL_DETAIL_PROTOTYPE.copy(
+      bookingInTime = bookingInTime,
+      movementReasonCode = movementReasonCode
+    )
+
+    val response = arrivalsService.confirmArrival(MOVE_ID, confirmArrivalDetail)
+
+    assertThat(response.offenderNo).isEqualTo(OFFENDER_NO)
+
+    verify { expectedOperation(OFFENDER_NO, confirmArrivalDetail) }
+    verify {
+      confirmedArrivalService.add(
+        MOVE_ID,
+        OFFENDER_NO,
+        PRISON_ID,
+        BOOKING_ID,
+        bookingInTime?.toLocalDate() ?: LocalDate.now(FIXED_CLOCK),
+        expectedArrivalType
       )
-    )
-    verify { prisonService.admitOffenderOnNewBooking(any(), any()) }
-    verify { confirmedArrivalService.add(any(), any(), any(), any(), any(), ArrivalType.NEW_BOOKING_EXISTING_OFFENDER) }
+    }
   }
 
   companion object {
@@ -184,20 +177,20 @@ class ArrivalsServiceConfirmArrivalTest {
     private val ZONE_ID: ZoneId = ZoneId.systemDefault()
     private val FIXED_CLOCK = Clock.fixed(FIXED_NOW, ZONE_ID)
 
-    val prototypeArrival = Arrival(
+    val ARRIVAL_PROTOTYPE = Arrival(
       id = MOVE_ID,
       firstName = FIRST_NAME,
       lastName = LAST_NAME,
       dateOfBirth = DATE_OF_BIRTH,
       prisonNumber = null,
       pncNumber = null,
-      date = LocalDate.now(),
+      date = LocalDate.now(FIXED_CLOCK),
       fromLocation = "Kingston-upon-Hull Crown Court",
       fromLocationType = LocationType.COURT,
       isCurrentPrisoner = false
     )
 
-    val confirmArrivalDetail = ConfirmArrivalDetail(
+    val CONFIRMED_ARRIVAL_DETAIL_PROTOTYPE = ConfirmArrivalDetail(
       firstName = FIRST_NAME,
       lastName = LAST_NAME,
       dateOfBirth = DATE_OF_BIRTH,
@@ -207,5 +200,12 @@ class ArrivalsServiceConfirmArrivalTest {
       imprisonmentStatus = "SENT03",
       bookingInTime = BOOKING_IN_TIME
     )
+
+    @JvmStatic
+    fun recallMovementReasonCodes(): Stream<String> = RECALL_MOVEMENT_REASON_CODES.stream()
+
+    @JvmStatic
+    fun nonRecallMovementReasonCodes(): Stream<String> =
+      (ALL_MOVEMENT_REASON_CODES - RECALL_MOVEMENT_REASON_CODES).stream()
   }
 }
